Changed paths: M       CHANGES.txt M       cg_tfce_estimate.m
(1) Added: Conversion to z-statistics according to palm which makes TFCE
    distribution more reliable for very small datasets.
(2) Changed: 3D data are now read with nifti strauture which is much more 
    faster.
(3) Changed: Estimation of uncorrected thresholds is not using cummulated
    histogram.
(4) Changed: Shared function for estimating GLM for F- and T-statistics.
(5) Added: Palm function for F-to-z transformation are now added.


git-svn-id: svn://141.35.69.214/tfce8@112 8e3dde7d-7981-41eb-959a-5c577105cc28